+++ crates/cli-support/src/interpreter/smoke_tests.rs
@@ -75,19 +75,20 @@ fn globals() {
             (import "__wbindgen_placeholder__" "__wbindgen_describe"
               (func $__wbindgen_describe (param i32)))
 
-            (global (mut i32) (i32.const 0))
+            (global $sp (mut i32) (i32.const 0))
 
             (func $foo
                 (local i32)
-                global.get 0
+                global.get $sp
                 local.set 0
                 local.get 0
                 call $__wbindgen_describe
                 local.get 0
-                global.set 0
+                global.set $sp
             )
 
             (export "foo" (func $foo))
+            (export "__stack_pointer" (global $sp))
         )
     "#;
     // __wbindgen_describe is called with a global - in Frame.eval we assume all access to globals is the stack pointer
@@ -145,7 +146,7 @@ fn loads_and_stores() {
             (import "__wbindgen_placeholder__" "__wbindgen_describe"
               (func $__wbindgen_describe (param i32)))
 
-            (global (mut i32) (i32.const 0))
+            (global $sp (mut i32) (i32.const 0))
             (memory 1)
 
             (func $foo
@@ -153,12 +154,12 @@ fn loads_and_stores() {
 
                 ;; decrement the stack pointer, setting our local to the
                 ;; lowest address of our stack
-                global.get 0
+                global.get $sp
                 i32.const 16
                 i32.sub
                 local.set 0
                 local.get 0
-                global.set 0
+                global.set $sp
 
                 ;; store 1 at fp+0
                 local.get 0
@@ -204,10 +205,11 @@ fn loads_and_stores() {
                 local.get 0
                 i32.const 16
                 i32.add
-                global.set 0
+                global.set $sp
             )
 
             (export "foo" (func $foo))
+            (export "__stack_pointer" (global $sp))
         )
     "#;
     interpret(wat, "foo", &[1, 50331650, 3]);
@@ -220,7 +222,7 @@ fn calling_functions() {
             (import "__wbindgen_placeholder__" "__wbindgen_describe"
               (func $__wbindgen_describe (param i32)))
 
-            (global i32 (i32.const 0))
+            (global $sp i32 (i32.const 0))
             (memory 1)
 
             (func $foo
@@ -233,6 +235,7 @@ fn calling_functions() {
             )
 
             (export "foo" (func $foo))
+            (export "__stack_pointer" (global $sp))
         )
     "#;
     interpret(wat, "foo", &[0]);
@@ -244,19 +247,19 @@ fn try_block() {
         (module
             (import "__wbindgen_placeholder__" "__wbindgen_describe"
               (func $__wbindgen_describe (param i32)))
-            (global (mut i32) (i32.const 0))
+            (global $sp (mut i32) (i32.const 0))
 
             (func $foo
                 (local i32)
 
                 ;; decrement the stack pointer, setting our local to the
                 ;; lowest address of our stack
-                global.get 0
+                global.get $sp
                 i32.const 16
                 i32.sub
                 local.set 0
                 local.get 0
-                global.set 0
+                global.set $sp
 
                 try
                     i32.const 1
@@ -268,10 +271,11 @@ fn try_block() {
                 local.get 0
                 i32.const 16
                 i32.add
-                global.set 0
+                global.set $sp
             )
 
             (export "foo" (func $foo))
+            (export "__stack_pointer" (global $sp))
         )
     "#;
     interpret(wat, "foo", &[1]);
@@ -283,19 +287,19 @@ fn try_table_block() {
         (module
             (import "__wbindgen_placeholder__" "__wbindgen_describe"
               (func $__wbindgen_describe (param i32)))
-            (global (mut i32) (i32.const 0))
+            (global $sp (mut i32) (i32.const 0))
 
             (func $foo
                 (local i32)
 
                 ;; decrement the stack pointer, setting our local to the
                 ;; lowest address of our stack
-                global.get 0
+                global.get $sp
                 i32.const 16
                 i32.sub
                 local.set 0
                 local.get 0
-                global.set 0
+                global.set $sp
 
                 (block $catch
                     (try_table (catch_all $catch)
@@ -308,10 +312,11 @@ fn try_table_block() {
                 local.get 0
                 i32.const 16
                 i32.add
-                global.set 0
+                global.set $sp
             )
 
             (export "foo" (func $foo))
+            (export "__stack_pointer" (global $sp))
         )
     "#;
     interpret(wat, "foo", &[1]);
@@ -324,7 +329,7 @@ fn calling_functions_with_args() {
             (import "__wbindgen_placeholder__" "__wbindgen_describe"
               (func $__wbindgen_describe (param i32)))
 
-            (global i32 (i32.const 0))
+            (global $sp i32 (i32.const 0))
             (memory 1)
 
             (func $foo
@@ -341,6 +346,7 @@ fn calling_functions_with_args() {
             )
 
             (export "foo" (func $foo))
+            (export "__stack_pointer" (global $sp))
         )
     "#;
     interpret(wat, "foo", &[1, 2]);
@@ -354,7 +360,7 @@ fn calling_function_with_args_out_of_order() {
             (import "__wbindgen_placeholder__" "__wbindgen_describe"
               (func $__wbindgen_describe (param i32)))
 
-            (global i32 (i32.const 0))
+            (global $sp i32 (i32.const 0))
             (memory 1)
 
             (func $foo
@@ -371,6 +377,7 @@ fn calling_function_with_args_out_of_order() {
             )
 
             (export "foo" (func $foo))
+            (export "__stack_pointer" (global $sp))
         )
     "#;
     interpret(wat, "foo", &[2, 1]);
@@ -383,7 +390,7 @@ fn blocks() {
             (import "__wbindgen_placeholder__" "__wbindgen_describe"
               (func $__wbindgen_describe (param i32)))
 
-            (global (mut i32) (i32.const 0))
+            (global $sp (mut i32) (i32.const 0))
             (memory 1)
 
             (func $foo
@@ -391,12 +398,12 @@ fn blocks() {
 
                 ;; decrement the stack pointer, setting our local to the
                 ;; lowest address of our stack
-                global.get 0
+                global.get $sp
                 i32.const 16
                 i32.sub
                 local.set 0
                 local.get 0
-                global.set 0
+                global.set $sp
 
                 (block
                     i32.const 0
@@ -407,10 +414,67 @@ fn blocks() {
                 local.get 0
                 i32.const 16
                 i32.add
-                global.set 0
+                global.set $sp
             )
             (export "foo" (func $foo))
+            (export "__stack_pointer" (global $sp))
         )
     "#;
     interpret(wat, "foo", &[0]);
+}
+
+// Test for issue #5080: interpreter should distinguish between __stack_pointer
+// and other globals like GOT.func.internal.*
+#[test]
+fn multiple_globals_with_stack_pointer() {
+    let wat = r#"
+        (module
+            (import "__wbindgen_placeholder__" "__wbindgen_describe"
+              (func $__wbindgen_describe (param i32)))
+
+            ;; __stack_pointer should be handled specially
+            (global $sp (mut i32) (i32.const 0))
+            ;; Other globals should be stored separately
+            (global $other1 i32 (i32.const 42))
+            (global $other2 (mut i32) (i32.const 100))
+
+            (func $foo
+                ;; Read other global - should get 42, not stack pointer
+                global.get $other1
+                call $__wbindgen_describe
+
+                ;; Read stack pointer - should get 32768
+                global.get $sp
+                call $__wbindgen_describe
+
+                ;; Modify other global
+                i32.const 200
+                global.set $other2
+
+                ;; Read modified other global
+                global.get $other2
+                call $__wbindgen_describe
+
+                ;; Modify stack pointer
+                global.get $sp
+                i32.const 16
+                i32.sub
+                global.set $sp
+
+                ;; Read stack pointer again - should get 32752
+                global.get $sp
+                call $__wbindgen_describe
+
+                ;; Restore stack pointer
+                global.get $sp
+                i32.const 16
+                i32.add
+                global.set $sp
+            )
+
+            (export "foo" (func $foo))
+            (export "__stack_pointer" (global $sp))
+        )
+    "#;
+    interpret(wat, "foo", &[42, 32768, 200, 32752]);
 }